+++ crates/cargo-zng/src/res/built_in.rs
@@ -170,7 +170,8 @@ macro_rules! built_in {
         ];
     };
 }
-built_in! { copy, glob, rp, sh, shf, warn, fail, apk, l10n }
+// TODO(breaking) support chaining tools without needing to repeat ".zr-"
+built_in! { copy, glob, rp, sh, shf, warn, fail, apk, l10n, sfx, sfxf }
 
 pub(crate) use l10n::release_langs;
 pub(crate) use sh::sh_run;