+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+* Add self-extracting executable builder tool for `cargo zng res`.
+    - Files with `.zr-sfx` or `.zr-sfxf` build a SFX.
+    - See `cargo zng res --tool sfx` for features.
+* Fix `zng-res::on-final=` not working when the request originates from a nested tool call.
+* On Windows, always use '/' separators in `cargo zng res` paths env vars.
 * Improve "unknown format" image error by including the first 24 bytes.
 * Fix renderer requesting texture ID from error images.